Fixing various warnings in glib/tests/sequence.c
In file included from glib/glibconfig.h:9,
                 from glib/gtypes.h:32,
                 from glib/galloca.h:32,
                 from glib/glib.h:30,
                 from glib/tests/sequence.c:2:
glib/tests/sequence.c: In function ‘check_integrity’:
glib/tests/sequence.c:139:27: error: comparison of integer expressions of different signedness: ‘int’ and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
   g_assert (info->n_items == g_queue_get_length (info->queue));
                           ^~
glib/gmacros.h:455:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
glib/tests/sequence.c:139:3: note: in expansion of macro ‘g_assert’
   g_assert (info->n_items == g_queue_get_length (info->queue));
   ^~~~~~~~
glib/tests/sequence.c:157:27: error: comparison of integer expressions of different signedness: ‘int’ and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
   g_assert (info->n_items == g_queue_get_length (info->queue));
                           ^~
glib/gmacros.h:455:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
glib/tests/sequence.c:157:3: note: in expansion of macro ‘g_assert’
   g_assert (info->n_items == g_queue_get_length (info->queue));
   ^~~~~~~~
glib/tests/sequence.c: In function ‘run_random_tests’:
glib/tests/sequence.c:554:55: error: comparison of integer expressions of different signedness: ‘guint’ {aka ‘unsigned int’} and ‘gint’ {aka ‘int’} [-Werror=sign-compare]
             g_assert (g_queue_get_length (seq->queue) == g_sequence_get_length (seq->sequence));
                                                       ^~
glib/gmacros.h:455:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
glib/tests/sequence.c:554:13: note: in expansion of macro ‘g_assert’
             g_assert (g_queue_get_length (seq->queue) == g_sequence_get_length (seq->sequence));
             ^~~~~~~~
glib/tests/sequence.c: In function ‘main’:
glib/tests/sequence.c:1404:17: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘long unsigned int’ [-Werror=sign-compare]
   for (i = 0; i < G_N_ELEMENTS (seeds); ++i)
                 ^

+++ glib/tests/sequence.c
@@ -136,7 +136,7 @@ check_integrity (SequenceInfo *info)
     g_printerr ("%d %d\n",
              g_sequence_get_length (info->sequence), info->n_items);
 #endif
-  g_assert (info->n_items == g_queue_get_length (info->queue));
+  g_assert (info->n_items == (gint) g_queue_get_length (info->queue));
   g_assert (g_sequence_get_length (info->sequence) == info->n_items);
 
   iter = g_sequence_get_begin_iter (info->sequence);
@@ -154,7 +154,7 @@ check_integrity (SequenceInfo *info)
       i++;
     }
 
-  g_assert (info->n_items == g_queue_get_length (info->queue));
+  g_assert (info->n_items == (gint) g_queue_get_length (info->queue));
   g_assert (g_sequence_get_length (info->sequence) == info->n_items);
 }
 
@@ -549,9 +549,9 @@ run_random_tests (gconstpointer d)
           break;
         case GET_ITER_AT_POS:
           {
-            int i;
+            guint i;
 
-            g_assert (g_queue_get_length (seq->queue) == g_sequence_get_length (seq->sequence));
+            g_assert (g_queue_get_length (seq->queue) == (guint) g_sequence_get_length (seq->sequence));
 
             for (i = 0; i < 10; ++i)
               {
@@ -1387,7 +1387,7 @@ int
 main (int argc,
       char **argv)
 {
-  gint i;
+  gsize i;
   guint32 seed;
   gchar *path;
